+++ pkg/module/offline/offline_test.go
@@ -43,15 +43,10 @@ func TestOffline_ArchiveOfflineMessage(t *testing.T) {
 	hostsMock := &hostsMock{}
 	hostsMock.IsLocalHostFunc = func(h string) bool { return h == "jackal.im" }
 
-	resManagerMock := &resourceManagerMock{}
-	resManagerMock.GetResourcesFunc = func(ctx context.Context, username string) ([]c2smodel.ResourceDesc, error) {
-		return nil, nil
-	}
 	hk := hook.NewHooks()
 	m := &Offline{
 		cfg:    Config{QueueSize: 100},
 		hosts:  hostsMock,
-		resMng: resManagerMock,
 		rep:    repMock,
 		hk:     hk,
 		logger: kitlog.NewNopLogger(),
@@ -70,10 +65,11 @@ func TestOffline_ArchiveOfflineMessage(t *testing.T) {
 	_ = m.Start(context.Background())
 	defer func() { _ = m.Stop(context.Background()) }()
 
-	_, _ = hk.Run(context.Background(), hook.C2SStreamWillRouteElement, &hook.ExecutionContext{
+	_, _ = hk.Run(hook.C2SStreamMessageRouted, &hook.ExecutionContext{
 		Info: &hook.C2SStreamInfo{
 			Element: msg,
 		},
+		Context: context.Background(),
 	})
 
 	// then
@@ -113,7 +109,6 @@ func TestOffline_ArchiveOfflineMessageQueueFull(t *testing.T) {
 		cfg:    Config{QueueSize: 100},
 		router: routerMock,
 		hosts:  hostsMock,
-		resMng: resManagerMock,
 		rep:    repMock,
 		hk:     hk,
 		logger: kitlog.NewNopLogger(),
@@ -132,10 +127,11 @@ func TestOffline_ArchiveOfflineMessageQueueFull(t *testing.T) {
 	_ = m.Start(context.Background())
 	defer func() { _ = m.Stop(context.Background()) }()
 
-	halted, err := hk.Run(context.Background(), hook.C2SStreamWillRouteElement, &hook.ExecutionContext{
+	halted, err := hk.Run(hook.C2SStreamMessageRouted, &hook.ExecutionContext{
 		Info: &hook.C2SStreamInfo{
 			Element: msg,
 		},
+		Context: context.Background(),
 	})
 
 	// then
@@ -152,11 +148,6 @@ func TestOffline_DeliverOfflineMessages(t *testing.T) {
 	// given
 	routerMock := &routerMock{}
 
-	output := bytes.NewBuffer(nil)
-	routerMock.RouteFunc = func(ctx context.Context, stanza stravaganza.Stanza) ([]jid.JID, error) {
-		_ = stanza.ToXML(output, true)
-		return nil, nil
-	}
 	hostsMock := &hostsMock{}
 	hostsMock.IsLocalHostFunc = func(h string) bool { return h == "jackal.im" }
 
@@ -184,6 +175,22 @@ func TestOffline_DeliverOfflineMessages(t *testing.T) {
 		return nil
 	}
 
+	stmMock := &c2sStreamMock{}
+	stmMock.UsernameFunc = func() string {
+		return "ortuman"
+	}
+	stmMock.InfoFunc = func() c2smodel.Info {
+		return c2smodel.NewInfoMap()
+	}
+
+	output := bytes.NewBuffer(nil)
+	stmMock.SendElementFunc = func(elem stravaganza.Element) <-chan error {
+		_ = elem.ToXML(output, true)
+		ch := make(chan error)
+		close(ch)
+		return ch
+	}
+
 	hk := hook.NewHooks()
 	m := &Offline{
 		cfg:    Config{QueueSize: 100},
@@ -202,10 +209,12 @@ func TestOffline_DeliverOfflineMessages(t *testing.T) {
 
 	pr := xmpputil.MakePresence(fromJID, toJID, stravaganza.AvailableType, nil)
 
-	_, _ = hk.Run(context.Background(), hook.C2SStreamPresenceReceived, &hook.ExecutionContext{
+	_, _ = hk.Run(hook.C2SStreamPresenceReceived, &hook.ExecutionContext{
 		Info: &hook.C2SStreamInfo{
 			Element: pr,
 		},
+		Sender:  stmMock,
+		Context: context.Background(),
 	})
 
 	// then